release 1.0

+++ app/src/main/java/com/sorcererxw/matthiasheiderichphotography/ui/activities/DetailActivity.java
@@ -1,32 +1,37 @@
-package com.sorcererxw.matthiasheidericphotography.ui.activities;
+package com.sorcererxw.matthiasheiderichphotography.ui.activities;
 
 import android.app.WallpaperManager;
+import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.drawable.Drawable;
+import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.support.annotation.NonNull;
+import android.support.design.widget.CoordinatorLayout;
+import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.graphics.Palette;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.ImageView;
 import android.widget.Toast;
 
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
-import com.bumptech.glide.load.resource.drawable.GlideDrawable;
-import com.bumptech.glide.request.RequestListener;
-import com.bumptech.glide.request.target.Target;
-import com.getbase.floatingactionbutton.FloatingActionButton;
-import com.getbase.floatingactionbutton.FloatingActionsMenu;
+import com.github.clans.fab.FloatingActionButton;
+import com.github.clans.fab.FloatingActionMenu;
+import com.sorcererxw.matthiasheiderichphotography.MHApplication;
+import com.sorcererxw.matthiasheiderichphotography.ui.views.TypefaceToolbar;
+import com.sorcererxw.matthiasheiderichphotography.util.DialogUtil;
+import com.sorcererxw.matthiasheiderichphotography.util.PermissionsHelper;
+import com.sorcererxw.matthiasheiderichphotography.util.ResourceUtil;
+import com.sorcererxw.matthiasheiderichphotography.util.StringUtil;
 import com.sorcererxw.matthiasheidericphotography.BuildConfig;
-import com.sorcererxw.matthiasheidericphotography.MHApplication;
 import com.sorcererxw.matthiasheidericphotography.R;
-import com.sorcererxw.matthiasheidericphotography.ui.views.TypefaceToolbar;
-import com.sorcererxw.matthiasheidericphotography.util.DialogUtil;
-import com.sorcererxw.matthiasheidericphotography.util.PermissionsHelper;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -44,13 +49,14 @@ import rx.schedulers.Schedulers;
 import uk.co.senab.photoview.PhotoView;
 
 public class DetailActivity extends AppCompatActivity {
-
+    @BindView(R.id.coordinatorLayout_detail)
+    CoordinatorLayout mCoordinatorLayout;
     @BindView(R.id.toolbar_detail)
     TypefaceToolbar mToolbar;
     @BindView(R.id.imageView_detail)
     PhotoView mImageView;
     @BindView(R.id.fab_detail)
-    FloatingActionsMenu mFAB;
+    FloatingActionMenu mFAB;
     @BindView(R.id.fab_detail_apply)
     FloatingActionButton mApplyFAB;
     @BindView(R.id.fab_detail_save)
@@ -73,7 +79,14 @@ public class DetailActivity extends AppCompatActivity {
     private MaterialDialog mSaveDialog;
     private MaterialDialog mSetDialog;
 
-    @Override
+    private Snackbar.Callback mSnackbarCallback = new Snackbar.Callback() {
+        @Override
+        public void onDismissed(Snackbar snackbar, int event) {
+            super.onDismissed(snackbar, event);
+            mFAB.setTranslationY(0);
+        }
+    };
+
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_detail);
@@ -86,12 +99,15 @@ public class DetailActivity extends AppCompatActivity {
         }
         mSaveDialog = DialogUtil.getProgressDialog(this, "Saving to local...");
         mSetDialog = DialogUtil.getProgressDialog(this, "Setting wallpaper...");
+
+        mFAB.hideMenuButton(false);
     }
 
     private void initImage() {
         mImageView.setScaleType(ImageView.ScaleType.CENTER_CROP);
         if (MHApplication.getInstance().getTmpDrawable() != null) {
             mImageView.setImageDrawable(MHApplication.getInstance().getTmpDrawable());
+            setFABColor();
         } else {
             Observable.just(mLink + "?format=1000w").map(new Func1<String, Drawable>() {
                 @Override
@@ -100,23 +116,6 @@ public class DetailActivity extends AppCompatActivity {
                         return Glide.with(DetailActivity.this)
                                 .load(s)
                                 .diskCacheStrategy(DiskCacheStrategy.ALL)
-                                .listener(new RequestListener<String, GlideDrawable>() {
-                                    @Override
-                                    public boolean onException(Exception e, String model,
-                                                               Target<GlideDrawable> target,
-                                                               boolean isFirstResource) {
-                                        return false;
-                                    }
-
-                                    @Override
-                                    public boolean onResourceReady(GlideDrawable resource,
-                                                                   String model,
-                                                                   Target<GlideDrawable> target,
-                                                                   boolean isFromMemoryCache,
-                                                                   boolean isFirstResource) {
-                                        return false;
-                                    }
-                                })
                                 .into(-1, -1)
                                 .get();
                     } catch (InterruptedException | ExecutionException e) {
@@ -132,6 +131,7 @@ public class DetailActivity extends AppCompatActivity {
                         public void call(Drawable drawable) {
                             if (drawable != null) {
                                 mImageView.setImageDrawable(drawable);
+                                setFABColor();
                             } else {
                                 finish();
                             }
@@ -181,10 +181,10 @@ public class DetailActivity extends AppCompatActivity {
             PermissionsHelper.requestWriteExternalStorage(this);
             return;
         }
-        if (!mSaveDialog.isShowing()) {
-            mSaveDialog.show();
-        }
         if (mRawBitmap == null) {
+            if (!mSaveDialog.isShowing()) {
+                mSaveDialog.show();
+            }
             download(new DownloadCallback() {
                 @Override
                 public void onFinish() {
@@ -196,15 +196,52 @@ public class DetailActivity extends AppCompatActivity {
             try {
                 File path = Environment.getExternalStoragePublicDirectory(
                         Environment.DIRECTORY_PICTURES);
-                File file = new File(path, "/Matthisa Heideric/" + "test.png");
-                out = new FileOutputStream(file);
-                mRawBitmap.compress(Bitmap.CompressFormat.PNG, 100, out);
-                Toast.makeText(this, "Success", Toast.LENGTH_SHORT).show();
+                File dir = new File(path, "/Matthisa Heideric");
+                if (!dir.exists()) {
+                    dir.mkdir();
+                }
+                final File file = new File(path, "/Matthisa Heideric/"
+                        + StringUtil.getFileNameFromLinkWithoutExtension(mLink) + ".png");
+                if (file.exists()) {
+                    Snackbar.make(mCoordinatorLayout, "Existed", Snackbar.LENGTH_LONG)
+                            .setAction("Open", new View.OnClickListener() {
+                                @Override
+                                public void onClick(View view) {
+                                    Intent intent = new Intent();
+                                    intent.setAction(Intent.ACTION_VIEW);
+                                    intent.setDataAndType(Uri.fromFile(file), "image/*");
+                                    startActivity(intent);
+                                }
+                            })
+                            .setActionTextColor(ResourceUtil.getColor(this, R.color.white))
+                            .setCallback(mSnackbarCallback)
+                            .show();
+                } else {
+                    if (!mSaveDialog.isShowing()) {
+                        mSaveDialog.show();
+                    }
+                    out = new FileOutputStream(file);
+                    mRawBitmap.compress(Bitmap.CompressFormat.PNG, 100, out);
+                    Snackbar.make(mCoordinatorLayout, "Success", Snackbar.LENGTH_LONG)
+                            .setAction("Open", new View.OnClickListener() {
+                                @Override
+                                public void onClick(View view) {
+                                    Intent intent = new Intent();
+                                    intent.setAction(Intent.ACTION_VIEW);
+                                    intent.setDataAndType(Uri.fromFile(file), "image/*");
+                                    startActivity(intent);
+                                }
+                            })
+                            .setActionTextColor(ResourceUtil.getColor(this, R.color.white))
+                            .setCallback(mSnackbarCallback)
+                            .show();
+                }
             } catch (Exception e) {
                 if (BuildConfig.DEBUG) {
                     e.printStackTrace();
                 }
-                Toast.makeText(this, "Fail", Toast.LENGTH_SHORT).show();
+                Snackbar.make(mCoordinatorLayout, "Fail", Snackbar.LENGTH_LONG)
+                        .setCallback(mSnackbarCallback).show();
             } finally {
                 try {
                     if (out != null) {
@@ -251,14 +288,16 @@ public class DetailActivity extends AppCompatActivity {
                     = WallpaperManager.getInstance(getApplicationContext());
             try {
                 wallpaperManager.setWallpaperOffsets(
-                        getWindow().getDecorView().getRootView().getWindowToken(), 1, 1);
+                        getWindow().getDecorView().getRootView().getWindowToken(), 2, 1);
                 wallpaperManager.setBitmap(mRawBitmap);
-                Toast.makeText(this, "Success", Toast.LENGTH_SHORT).show();
+                Snackbar.make(mCoordinatorLayout, "Success", Snackbar.LENGTH_LONG)
+                        .setCallback(mSnackbarCallback).show();
             } catch (IOException e) {
                 if (BuildConfig.DEBUG) {
                     e.printStackTrace();
                 }
-                Toast.makeText(this, "Fail", Toast.LENGTH_SHORT).show();
+                Snackbar.make(mCoordinatorLayout, "Fail", Snackbar.LENGTH_LONG)
+                        .setCallback(mSnackbarCallback).show();
             }
             mSetDialog.dismiss();
         }
@@ -272,5 +311,61 @@ public class DetailActivity extends AppCompatActivity {
             super.onBackPressed();
         }
         return false;
+    }
+
+
+    private void setFABColor() {
+        Observable.just(mLink + "?format=1000w").map(new Func1<String, Palette.Swatch>() {
+            @Override
+            public Palette.Swatch call(String s) {
+                Bitmap bitmap = null;
+                try {
+                    bitmap = Glide.with(DetailActivity.this)
+                            .load(s)
+                            .asBitmap()
+                            .diskCacheStrategy(DiskCacheStrategy.ALL)
+                            .into(-1, -1)
+                            .get();
+                } catch (InterruptedException | ExecutionException e) {
+                    e.printStackTrace();
+                }
+                if (bitmap != null) {
+                    Palette palette = Palette.from(bitmap).generate();
+                    Palette.Swatch[] swatches = new Palette.Swatch[]{
+                            palette.getMutedSwatch(),
+                            palette.getLightMutedSwatch(),
+                            palette.getDarkMutedSwatch(),
+                            palette.getVibrantSwatch(),
+                            palette.getLightVibrantSwatch(),
+                            palette.getDarkVibrantSwatch(),
+                    };
+                    for (Palette.Swatch sc : swatches) {
+                        if (sc != null) {
+                            return sc;
+                        }
+                    }
+                    return null;
+                }
+                return null;
+            }
+        })
+                .subscribeOn(Schedulers.newThread())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(new Action1<Palette.Swatch>() {
+                    @Override
+                    public void call(Palette.Swatch swatch) {
+                        int color = ResourceUtil.getColor(DetailActivity.this, R.color.accent);
+                        if (swatch != null) {
+                            color = swatch.getRgb();
+                        }
+                        mFAB.setMenuButtonColorNormal(color);
+                        mFAB.setMenuButtonColorPressed(color);
+                        mApplyFAB.setColorNormal(color);
+                        mApplyFAB.setColorPressed(color);
+                        mSaveFAB.setColorNormal(color);
+                        mSaveFAB.setColorPressed(color);
+                        mFAB.showMenuButton(true);
+                    }
+                });
     }
 }
